+++ src/Mod/CAM/PathSimulator/AppGL/TextureLoader.cpp
@@ -55,8 +55,7 @@ TextureLoader::TextureLoader(std::string imgFolder,
                              int textureSize)
     : mImageFolder(imgFolder)
 {
-    size_t buffsize =
-        static_cast<size_t>(textureSize) * static_cast<size_t>(textureSize) * sizeof(unsigned int);
+    int buffsize = textureSize * textureSize * sizeof(unsigned int);
     mRawData = (unsigned int*)malloc(buffsize);
     if (mRawData == nullptr) {
         return;
